Python/FlashFrame-1/requirements.txt
Pillow==9.0.1
tkinter==0.1.0